+++ packages/pluggableWidgets/datagrid-web/src/components/WidgetFooter.tsx
@@ -33,3 +33,17 @@ export function WidgetFooter(props: WidgetFooterProps): ReactElement | null {
         </div>
     );
 }
+
+const SelectionCounter = observer(function SelectionCounter() {
+    const { selectionCountStore, selectActionHelper } = useDatagridRootScope();
+
+    return (
+        <If condition={selectionCountStore.selectedCountText !== ""}>
+            <span className="widget-datagrid-selection-count">{selectionCountStore.selectedCountText}</span>
+            &nbsp;
+            <button className="widget-datagrid-btn-link" onClick={selectActionHelper.onClearSelection}>
+                {selectionCountStore.clearSelectionText}
+            </button>
+        </If>
+    );
+});