+++ packages/pluggableWidgets/datagrid-web/typings/DatagridProps.d.ts
@@ -7,14 +7,6 @@ import { ComponentType, CSSProperties, ReactNode } from "react";
 import { ActionValue, DynamicValue, EditableValue, ListValue, ListActionValue, ListAttributeValue, ListAttributeListValue, ListExpressionValue, ListWidgetValue, SelectionSingleValue, SelectionMultiValue } from "mendix";
 import { Big } from "big.js";
 
-export type ItemSelectionMethodEnum = "checkbox" | "rowClick";
-
-export type ItemSelectionModeEnum = "toggle" | "clear";
-
-export type SelectionCountPositionEnum = "top" | "bottom" | "off";
-
-export type LoadingTypeEnum = "spinner" | "skeleton";
-
 export type ShowContentAsEnum = "attribute" | "dynamicText" | "customContent";
 
 export type HidableEnum = "yes" | "hidden" | "no";
@@ -49,6 +41,14 @@ export interface ColumnsType {
     wrapText: boolean;
 }
 
+export type OnClickTriggerEnum = "single" | "double";
+
+export type ItemSelectionMethodEnum = "checkbox" | "rowClick";
+
+export type ItemSelectionModeEnum = "toggle" | "clear";
+
+export type LoadingTypeEnum = "spinner" | "skeleton";
+
 export type PaginationEnum = "buttons" | "virtualScrolling" | "loadMore";
 
 export type ShowPagingButtonsEnum = "always" | "auto";
@@ -56,8 +56,6 @@ export type ShowPagingButtonsEnum = "always" | "auto";
 export type PagingPositionEnum = "bottom" | "top" | "both";
 
 export type ShowEmptyPlaceholderEnum = "none" | "custom";
-
-export type OnClickTriggerEnum = "single" | "double";
 
 export type ConfigurationStorageTypeEnum = "attribute" | "localStorage";
 
@@ -90,20 +88,37 @@ export interface DatagridContainerProps {
     class: string;
     style?: CSSProperties;
     tabIndex?: number;
-    advanced: boolean;
     datasource: ListValue;
     refreshInterval: number;
+    columns: ColumnsType[];
+    columnsFilterable: boolean;
+    onClickTrigger: OnClickTriggerEnum;
+    onClick?: ListActionValue;
+    onSelectionChange?: ActionValue;
+    filtersPlaceholder?: ReactNode;
     itemSelection?: SelectionSingleValue | SelectionMultiValue;
     itemSelectionMethod: ItemSelectionMethodEnum;
     itemSelectionMode: ItemSelectionModeEnum;
     showSelectAllToggle: boolean;
     keepSelection: boolean;
+<<<<<<< HEAD
+<<<<<<< HEAD
     selectionCountPosition: SelectionCountPositionEnum;
     clearSelectionButtonLabel?: DynamicValue<string>;
+=======
+    selectAllPagesEnabled: boolean;
+    selectAllPagesPageSize: number;
+<<<<<<< HEAD
+    selectingAllLabel?: DynamicValue<string>;
+    cancelSelectionLabel?: DynamicValue<string>;
+>>>>>>> 2e4671e66 (feat(datagrid-web): add multipage selection to dg2)
+=======
+>>>>>>> 448ce2ee2 (feat: add sab vm)
+=======
+    enableSelectAll: boolean;
+>>>>>>> 298cb49db (refactor: change texts & etc)
     loadingType: LoadingTypeEnum;
     refreshIndicator: boolean;
-    columns: ColumnsType[];
-    columnsFilterable: boolean;
     pageSize: number;
     pagination: PaginationEnum;
     showPagingButtons: ShowPagingButtonsEnum;
@@ -113,10 +128,6 @@ export interface DatagridContainerProps {
     showEmptyPlaceholder: ShowEmptyPlaceholderEnum;
     emptyPlaceholder?: ReactNode;
     rowClass?: ListExpressionValue<string>;
-    onClickTrigger: OnClickTriggerEnum;
-    onClick?: ListActionValue;
-    onSelectionChange?: ActionValue;
-    filtersPlaceholder?: ReactNode;
     columnsSortable: boolean;
     columnsResizable: boolean;
     columnsDraggable: boolean;
@@ -129,8 +140,14 @@ export interface DatagridContainerProps {
     cancelExportLabel?: DynamicValue<string>;
     selectRowLabel?: DynamicValue<string>;
     selectAllRowsLabel?: DynamicValue<string>;
+    selectingAllLabel?: DynamicValue<string>;
+    cancelSelectionLabel?: DynamicValue<string>;
     selectedCountTemplateSingular?: DynamicValue<string>;
     selectedCountTemplatePlural?: DynamicValue<string>;
+    selectAllText: DynamicValue<string>;
+    selectAllTemplate: DynamicValue<string>;
+    allSelectedText: DynamicValue<string>;
+    clearSelectionCaption: DynamicValue<string>;
 }
 
 export interface DatagridPreviewProps {
@@ -144,20 +161,36 @@ export interface DatagridPreviewProps {
     readOnly: boolean;
     renderMode: "design" | "xray" | "structure";
     translate: (text: string) => string;
-    advanced: boolean;
     datasource: {} | { caption: string } | { type: string } | null;
     refreshInterval: number | null;
+    columns: ColumnsPreviewType[];
+    columnsFilterable: boolean;
+    onClickTrigger: OnClickTriggerEnum;
+    onClick: {} | null;
+    onSelectionChange: {} | null;
+    filtersPlaceholder: { widgetCount: number; renderer: ComponentType<{ children: ReactNode; caption?: string }> };
     itemSelection: "None" | "Single" | "Multi";
     itemSelectionMethod: ItemSelectionMethodEnum;
     itemSelectionMode: ItemSelectionModeEnum;
     showSelectAllToggle: boolean;
     keepSelection: boolean;
+<<<<<<< HEAD
+
     selectionCountPosition: SelectionCountPositionEnum;
     clearSelectionButtonLabel: string;
+
+    selectAllPagesEnabled: boolean;
+    selectAllPagesPageSize: number | null;
+
+    selectingAllLabel: string;
+    cancelSelectionLabel: string;
+
+
+=======
+    enableSelectAll: boolean;
+>>>>>>> 298cb49db (refactor: change texts & etc)
     loadingType: LoadingTypeEnum;
     refreshIndicator: boolean;
-    columns: ColumnsPreviewType[];
-    columnsFilterable: boolean;
     pageSize: number | null;
     pagination: PaginationEnum;
     showPagingButtons: ShowPagingButtonsEnum;
@@ -167,10 +200,6 @@ export interface DatagridPreviewProps {
     showEmptyPlaceholder: ShowEmptyPlaceholderEnum;
     emptyPlaceholder: { widgetCount: number; renderer: ComponentType<{ children: ReactNode; caption?: string }> };
     rowClass: string;
-    onClickTrigger: OnClickTriggerEnum;
-    onClick: {} | null;
-    onSelectionChange: {} | null;
-    filtersPlaceholder: { widgetCount: number; renderer: ComponentType<{ children: ReactNode; caption?: string }> };
     columnsSortable: boolean;
     columnsResizable: boolean;
     columnsDraggable: boolean;
@@ -184,6 +213,12 @@ export interface DatagridPreviewProps {
     cancelExportLabel: string;
     selectRowLabel: string;
     selectAllRowsLabel: string;
+    selectingAllLabel: string;
+    cancelSelectionLabel: string;
     selectedCountTemplateSingular: string;
     selectedCountTemplatePlural: string;
+    selectAllText: string;
+    selectAllTemplate: string;
+    allSelectedText: string;
+    clearSelectionCaption: string;
 }