+++ packages/pluggableWidgets/gallery-web/typings/GalleryProps.d.ts
@@ -61,6 +61,7 @@ export interface GalleryContainerProps {
     ariaLabelItem?: ListExpressionValue<string>;
     selectedCountTemplateSingular?: DynamicValue<string>;
     selectedCountTemplatePlural?: DynamicValue<string>;
+    clearSelectionCaption: DynamicValue<string>;
 }
 
 export interface GalleryPreviewProps {
@@ -109,4 +110,5 @@ export interface GalleryPreviewProps {
     ariaLabelItem: string;
     selectedCountTemplateSingular: string;
     selectedCountTemplatePlural: string;
+    clearSelectionCaption: string;
 }